// src/types/response.ts
export interface ApiResponse {
  currency_code: string;
  distance_to_cc: number;
  hotel_include_breakfast: 1 | 0;
  min_total_price: number;
  latitude: number;
  longitude: number;
  hotel_name_trans: string;
  currencycode: string; //PLN
  address: string; // Mielczarskiego 16/2C
  url: string;
  is_free_cancellable: 1 | 0;
  distance: number; // 1.67
  review_score: number; // 8.32
}